clean up config code structure
