+++ webapp/src/components/auth/sign-up.tsx
@@ -28,7 +28,7 @@ import {
   InputGroupInput,
 } from "@/components/ui/input-group"
 import { Spinner } from "@/components/ui/spinner"
-import { usePublicConfig } from "@/lib/config"
+import { usePublicConfig } from "@/components/public-config-provider"
 import { cn } from "@/lib/utils"
 import { PasswordStrengthMeter } from "./password-strength-meter"
 import { ProviderButtons, type SocialLayout } from "./provider-buttons"
