+++ webapp/src/routes/configure/-lib/types.ts
@@ -1,4 +1,4 @@
-import type { ConfigOption } from "@/lib/config"
+import type { ConfigOption } from "@/lib/types"
 
 export interface ConfigureField {
   key: string
